# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Inventario",
        "description": "API para gestionar el inventario de items",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/v1",
            "description": "Servidor local"
        }
    ],
    "paths": {
        "/items/{sku}/inventory": {
            "get": {
                "tags": [
                    "Inventories"
                ],
                "summary": "Lista el inventario de un item por su SKU",
                "description": "Obtiene el inventario de un item especificado por su SKU. Incluye detalles de color, tamaño y precio.",
                "operationId": "getItemInventory",
                "parameters": [
                    {
                        "name": "sku",
                        "in": "path",
                        "description": "SKU del item a buscar",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Inventario encontrado"
                    },
                    "404": {
                        "description": "Item no encontrado"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            },
            "post": {
                "tags": [
                    "Inventories"
                ],
                "summary": "Crea un nuevo inventario para un item",
                "description": "Crea un nuevo inventario para un item especificado por su SKU. Requiere color_id, size_id y price.",
                "operationId": "createInventory",
                "parameters": [
                    {
                        "name": "sku",
                        "in": "path",
                        "description": "SKU del item",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "color_id",
                                    "size_id",
                                    "price"
                                ],
                                "properties": {
                                    "color_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "size_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "price": {
                                        "type": "integer",
                                        "example": 15000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Inventario creado"
                    },
                    "404": {
                        "description": "Item no encontrado"
                    },
                    "500": {
                        "description": "Error al crear el inventario"
                    }
                }
            }
        },
        "/items/{sku}/inventory/{inventory_id}": {
            "get": {
                "tags": [
                    "Inventories"
                ],
                "summary": "Obtiene un inventario por ID y SKU del item",
                "operationId": "getInventoryBySkuAndId",
                "parameters": [
                    {
                        "name": "sku",
                        "in": "path",
                        "description": "SKU del item",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "inventory_id",
                        "in": "path",
                        "description": "ID del inventario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Inventario encontrado"
                    },
                    "404": {
                        "description": "Inventario o item no encontrado"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            },
            "put": {
                "tags": [
                    "Inventories"
                ],
                "summary": "Actualiza un inventario por ID y SKU del item",
                "operationId": "updateInventory",
                "parameters": [
                    {
                        "name": "sku",
                        "in": "path",
                        "description": "SKU del item",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "inventory_id",
                        "in": "path",
                        "description": "ID del inventario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "color_id",
                                    "size_id",
                                    "price"
                                ],
                                "properties": {
                                    "color_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "size_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "price": {
                                        "type": "number",
                                        "example": 15000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Inventario actualizado"
                    },
                    "404": {
                        "description": "Inventario o item no encontrado"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Inventories"
                ],
                "summary": "Elimina el inventario de un item por su SKU y su id de inventario",
                "description": "Elimina el inventario de un item especificado por su SKU e identificador de tabla. Elimina todas las entradas de inventario asociadas a ese SKU.",
                "operationId": "deleteItemInventoryBySku",
                "parameters": [
                    {
                        "name": "sku",
                        "in": "path",
                        "description": "SKU del item a eliminar",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "inventory_id",
                        "in": "path",
                        "description": "ID del inventario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Inventario eliminado"
                    },
                    "404": {
                        "description": "Item no encontrado"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    },
                    "204": {
                        "description": "Registro eliminado exitosamente"
                    }
                }
            }
        },
        "/items": {
            "get": {
                "tags": [
                    "Items"
                ],
                "summary": "Lista todos los items en el inventario",
                "operationId": "getItems",
                "responses": {
                    "200": {
                        "description": "Lista de items en el inventario"
                    },
                    "404": {
                        "description": "No se encontraron items"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            }
        },
        "/items/{sku}": {
            "get": {
                "tags": [
                    "Items"
                ],
                "summary": "Obtiene un item por su SKU",
                "operationId": "getItemBySku",
                "parameters": [
                    {
                        "name": "sku",
                        "in": "path",
                        "description": "SKU del item a buscar",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Item encontrado"
                    },
                    "404": {
                        "description": "Item no encontrado"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            },
            "put": {
                "tags": [
                    "Items"
                ],
                "summary": "Actualiza un item por su SKU",
                "operationId": "updateItem",
                "parameters": [
                    {
                        "name": "sku",
                        "in": "path",
                        "description": "SKU del item a actualizar",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Item"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Item actualizado exitosamente"
                    },
                    "404": {
                        "description": "Item no encontrado"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Items"
                ],
                "summary": "Elimina un item por su SKU",
                "operationId": "deleteItem",
                "parameters": [
                    {
                        "name": "sku",
                        "in": "path",
                        "description": "SKU del item a eliminar",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Item no encontrado"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    },
                    "204": {
                        "description": "Registro eliminado exitosamente"
                    }
                }
            }
        },
        "/items/": {
            "post": {
                "tags": [
                    "Items"
                ],
                "summary": "Crea un nuevo item",
                "operationId": "createItem",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Item"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Item creado exitosamente"
                    },
                    "400": {
                        "description": "Solicitud inválida"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Inventory": {
                "required": [
                    "color_id",
                    "size_id",
                    "price"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "color_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "size_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "price": {
                        "type": "integer",
                        "example": 15000
                    },
                    "color": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 1
                            },
                            "name": {
                                "type": "string",
                                "example": "Rojo"
                            }
                        },
                        "type": "object"
                    },
                    "size": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 2
                            },
                            "name": {
                                "type": "string",
                                "example": "M"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "Item": {
                "required": [
                    "sku",
                    "title",
                    "status"
                ],
                "properties": {
                    "sku": {
                        "type": "string",
                        "example": "SKU1234"
                    },
                    "title": {
                        "type": "string",
                        "example": "Zapato Deportivo"
                    },
                    "description": {
                        "type": "string",
                        "example": "Zapato para correr",
                        "nullable": true
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "active",
                            "inactive"
                        ],
                        "example": "active"
                    },
                    "inventories": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Inventory"
                        }
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Inventories",
            "description": "Operaciones relacionadas con productos en el stock"
        },
        {
            "name": "Items",
            "description": "Operaciones relacionadas con los items del inventario"
        }
    ]
}